api/requirements.txt
# DeepWiki-Open API 依赖

# Web框架
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
pydantic>=2.0.0

# AI模型客户端
google-generativeai>=0.3.0
openai>=1.76.2
adalflow>=0.1.0
zhipuai>=1.0.0

# 数据处理
tiktoken>=0.5.0
numpy>=1.24.0
faiss-cpu>=1.7.4
langid>=1.1.6
jinja2>=3.1.2

# HTTP和网络
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.4
websockets>=11.0.3
backoff>=2.0.0

# 配置和环境
python-dotenv>=1.0.0

# 可选依赖
ollama>=0.4.8

# 云服务
boto3>=1.34.0
azure-identity>=1.12.0
azure-core>=1.24.0